requirements.txt
fastapi 
uvicorn 
opencv-python 
google-genai 
pillow
python-dotenv
picamera2 
